APIGEN-822 Add android tests for hello world arrays
Tests for nested arrays and arrays containing all
franca data types are added to test directory in
separate commit.
The current support for byte arrays is not
overwritten and still will result in byte[] and
not in List<Byte>. The usual use case of large byte
arrays would have a decreased performance if every
byte is boxed in a Byte wrapper object and added to
the list. Therefore the example with UInt8 is
replaced.

+++ libhello/src/hello/HelloWorldArrays.cpp
@@ -14,16 +14,17 @@
 
 namespace hello
 {
+
 ::std::vector< ::std::string >
 HelloWorldArrays::methodWithArray( const ::std::vector< ::std::string >& input )
 {
-    return {input.rbegin( ), input.rend( )};
+    return { input.rbegin( ), input.rend( ) };
 }
 
-::std::vector< uint8_t >
-HelloWorldArrays::methodWithArrayInline( const ::std::vector< uint8_t >& input )
+::std::vector<  uint32_t >
+HelloWorldArrays::methodWithArrayInline( const ::std::vector<  uint32_t >& input )
 {
-    return {input.rbegin( ), input.rend( )};
+    return { input.rbegin( ), input.rend( ) };
 }
 
 ::std::vector< ::hello::HelloWorldArrays::ExampleStruct >
@@ -32,4 +33,5 @@ HelloWorldArrays::methodWithStructArray(
 {
     return {input.rbegin( ), input.rend( )};
 }
+
 }
